+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v3.19.0-dev (May 07, 2026)
+- Add test to catch Windows babel-loader exclude regex bug [#3840](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3840)
+
 ## v3.18.0 (May 07, 2026)
 - Add option to keep original User Agent header in proxy requests [#3798](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3798)
 - WIP: Add support for HttpOnly session cookies [#3804](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3804)